Main structure of the API code
Basic outline for how the telescope code will run and allow for runtime tracking is proposed. The idea is an object oriented approach, where we use an instance of a Telescope object to control static variables of the system. Tracking is handled by passing an arbitrary tracking function to the Telescope, which calls this function every fixed time intervals to update the target and point to this updated location. Further details are commented in the code.