Completion of Telescope API (sans Encoders)
Finished the activeTrack function and changed the variables to work within (0,360) azimuthal angle coordinate system. The Telescope API still needs the encoder system and ability to update the current angle stored in the Telescope object. It may also need error correction to account for lost steps, if lost steps prove to be an issue. However, all of these improvements are easy to implement.